docs(core): update info label examples
