--- libs/docs/core/info-label/examples/info-label-default-example.component.html
@@ -1,1 +1,0 @@
-<fd-info-label label="Default label" title="Default Label"></fd-info-label>
