+++ libs/docs/core/info-label/examples/info-label-example.component.ts
@@ -1,0 +1,15 @@
+import { Component } from '@angular/core';
+import { InfoLabelColorInput, InfoLabelModule } from '@fundamental-ngx/core/info-label';
+
+const infoLabelColors: InfoLabelColorInput[] = ['1', '2', '3', '4', '5', '6', '7', '8', '9', '10'];
+
+@Component({
+    selector: 'fd-info-label-example',
+    templateUrl: './info-label-example.component.html',
+    styleUrls: ['./info-label-example.component.scss'],
+    standalone: true,
+    imports: [InfoLabelModule]
+})
+export class InfoLabelExampleComponent {
+    infoLabelColors = infoLabelColors;
+}
