--- libs/docs/core/info-label/examples/info-label-icon-numeric-example.component.html
@@ -1,15 +1,0 @@
-<fd-info-label type="numeric" color="1" label="6" title="Numeric Label 6" [style.margin-right.px]="20"></fd-info-label>
-<fd-info-label
-    type="numeric"
-    color="2"
-    label="42k"
-    title="Numeric Label 42k"
-    [style.margin-right.px]="20"
-></fd-info-label>
-<fd-info-label
-    type="numeric"
-    color="3"
-    label="14.7"
-    title="Numeric Label 14.7"
-    [style.margin-right.px]="20"
-></fd-info-label>
