--- libs/docs/core/info-label/examples/info-label-text-example.component.html
@@ -1,5 +1,0 @@
-@for (color of infoLabelColors; track color) {
-    <div [style.margin-top.px]="10">
-        <fd-info-label label="Info Label" [color]="color" title="Info Label"></fd-info-label>
-    </div>
-}
