--- libs/docs/core/info-label/examples/info-label-text-icon-example.component.html
@@ -1,43 +1,0 @@
-<div [style.display]="'flex'">
-    <div [style.margin]="'0 10px'">
-        @for (item of infoLabelColors; track item) {
-            <div [style.margin-top.px]="10">
-                <fd-info-label
-                    type="icon"
-                    glyph="hide"
-                    label="Info label"
-                    [color]="item"
-                    title="Info Label with Icon"
-                ></fd-info-label>
-            </div>
-        }
-    </div>
-    <div [style.margin]="'0 10px'">
-        @for (item of infoLabelColors; track item) {
-            <div [style.margin-top.px]="10">
-                <fd-info-label
-                    type="icon"
-                    glyph="component"
-                    font="BusinessSuiteInAppSymbols"
-                    label="Info label"
-                    [color]="item"
-                    title="Info Label with Icon"
-                ></fd-info-label>
-            </div>
-        }
-    </div>
-    <div [style.margin]="'0 10px'">
-        @for (item of infoLabelColors; track item) {
-            <div [style.margin-top.px]="10">
-                <fd-info-label
-                    type="icon"
-                    glyph="technicalsystem"
-                    font="SAP-icons-TNT"
-                    label="Info label"
-                    [color]="item"
-                    title="Info Label with Icon"
-                ></fd-info-label>
-            </div>
-        }
-    </div>
-</div>
