+++ libs/docs/core/info-label/info-label-docs.component.html
@@ -1,44 +1,13 @@
-<fd-docs-section-title id="default" componentName="info-label"> Default Info Label </fd-docs-section-title>
-<component-example>
-    <fd-info-label-default-example></fd-info-label-default-example>
-</component-example>
-<code-example [exampleFiles]="defaultInfoLabelHtmlType"></code-example>
-
-<separator></separator>
-
-<fd-docs-section-title id="text" componentName="info-label"> Info Label With Text </fd-docs-section-title>
+<fd-docs-section-title id="default" componentName="info-label">
+    Info Label Examples(Text, Icon, Number)</fd-docs-section-title
+>
 <description
     >To apply semantic color to the Info Label use the <code>[color]</code> property. Available options are numbers from
-    1 to 10.</description
+    1 to 10.<br />Set the <code>[type]</code> property to 'icon' for Info Label with text and icon.<br />Set the
+    <code>[type]</code> property to 'numeric' for Info Label with a number. For Info Label with icon only set the
+    property to 'icon'.</description
 >
 <component-example>
-    <fd-info-label-text-example></fd-info-label-text-example>
+    <fd-info-label-example></fd-info-label-example>
 </component-example>
-<code-example [exampleFiles]="InfoLabelTextExample"></code-example>
-
-<separator></separator>
-
-<fd-docs-section-title id="text-and-icon" componentName="info-label">
-    Info Label With Text and Icon
-</fd-docs-section-title>
-<description>Set the <code>[type]</code> property to 'icon' for Info Label with text and icon. </description>
-<component-example>
-    <fd-info-label-text-icon-example></fd-info-label-text-icon-example>
-</component-example>
-<code-example [exampleFiles]="InfoLabelTextIconExample"></code-example>
-
-<separator></separator>
-
-<fd-docs-section-title id="numeric" componentName="info-label">
-    Info Label With a Number or an Icon
-</fd-docs-section-title>
-<description
-    >Set the <code>[type]</code> property to 'numeric' for Info Label with a number. For Info Label with icon only set
-    the property to 'icon'.
-</description>
-<component-example>
-    <fd-info-label-numeric-icon-example></fd-info-label-numeric-icon-example>
-</component-example>
-<code-example [exampleFiles]="InfoLableNumericIconExample"></code-example>
-
-<separator></separator>
+<code-example [exampleFiles]="defaultInfoLabelHtmlType"></code-example>
